+++ packages/cli/command/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @mash-up-web-toolkit/command
 
+## 0.0.20
+
+### Patch Changes
+
+- Updated dependencies [04ceca8]
+  - @mash-up-web-toolkit/generate-api-config@0.0.3
+
+## 0.0.19
+
+### Patch Changes
+
+- api 제너레이터 moduleNameFirstTag 속성 추가
+- Updated dependencies
+  - @mash-up-web-toolkit/generate-api@0.0.15
+
 ## 0.0.18
 
 ### Patch Changes